# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Laravel Jwt auth'
  description: 'A simple REST API in Laravel'
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/profile:
    post:
      summary: Profile
      operationId: profile
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer <Token>'
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            text/plain:
              schema:
                type: string
                example: "{\n     'id' => \"1\",\n     'name' => 'wael',\n    'email' => 'wael@gmail.com'\n}"
      tags:
        - Products
      security: []
...